Move All_FILTERS to filter_order_and_default_pipelines
